additional minor changes

+++ SupportDesk.Application/Handlers/Commands/DeleteTicketCommandHandler.cs
@@ -1,5 +1,8 @@
 ﻿using MediatR;
 using SupportDesk.Application.Commands.Tickets;
+using SupportDesk.Core.Enums;
+using SupportDesk.Core.Errors;
+using SupportDesk.Core.Exceptions;
 using SupportDesk.Core.Repositories;
 
 namespace SupportDesk.Application.Handlers.Commands;
@@ -18,6 +21,9 @@ public class DeleteTicketCommandHandler : IRequestHandler<DeleteTicketCommand, b
 		var ticket = await _ticketRepository.GetTrackedTicketById(request.Id, cancellationToken);
 		if (ticket is null)
 			return false;
+
+		if (ticket.Status == TicketStatus.Closed)
+			throw new BusinessRuleException(TicketErrors.Closed);
 
 		await _ticketRepository.DeleteTicket(ticket, cancellationToken);
 		return true;
